// backup/restore-point-2025-12-12-dex-trades/bzr-frontend/src/types/api.ts
export interface InternalTransfer {
  hash: string;
  blockNumber: string;
  timeStamp: string;
  from: string;
  to: string;
  value: string;
  contractAddress?: string;
  input?: string;
  type?: string;
  isError?: string;
  errCode?: string;
  chainName: string;
  chainId: number;
}